feat(11-01): add multi-stage Dockerfile for MCP server
- Multi-stage build: builder + runtime stages
- Base: python:3.11-slim-bookworm
- Non-root user (mcpuser) for security
- Health check using Python urllib at /health
- Port 8080 exposed for AWS ALB

Standalone mode support:
- auth_helpers.py: Pure utility functions without DB dependencies
- transport.py: Conditionally loads infrastructure middleware
- server.py: Uses auth_helpers instead of auth for header extraction
- Image size: ~353MB (acceptable for Python + deps)
